+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.4] - 2025-02-04
+
 ### Added
 
 - Added new IncludeRequired parameter for the TypeValue check. This because at TypeValue level,